update motor to use floats and SDL_bools

+++ src/mx_motor.h
@@ -5,7 +5,7 @@
 
 MX_Motor_Handle create_motor(void);
 void destroy_motor(MX_Motor_Handle motor_pointer);
-void run_game_loop(MX_Hull_Handle hull_pointer, void (*update)(Uint64 delta_ms));
+void run_game_loop(MX_Hull_Handle hull_pointer, void (*update)(float delta_ms));
 void toggle_pause(MX_Hull_Handle hull_pointer);
 void end_game_loop(MX_Hull_Handle hull_pointer);
 Uint64 is_running(MX_Hull_Handle hull_pointer);
